vesys/06_WebServices2/src/ch/fhnw/ds/graph/server/GraphService.java
package ch.fhnw.ds.graph.server;

import java.util.LinkedList;
import java.util.List;

import javax.jws.WebService;
import javax.jws.soap.SOAPBinding;
import javax.jws.soap.SOAPBinding.Style;
import javax.jws.soap.SOAPBinding.Use;

@WebService
@SOAPBinding(style=Style.DOCUMENT, use=Use.LITERAL)
public interface GraphService {
	public boolean isTree(Node root);
	public boolean isDAG(Node root);
	
	static class Node {
		private List<Node> children = new LinkedList<Node>();
		private String value;

		public String getValue() {
			return value;
		}
		public void setValue(String value) {
			this.value = value;
		}

		public List<Node> getChildren() {
			return children;
		}
		public void setChildren(List<Node> children) {
			this.children = children;
		}
	}
}
